refactor(82-01): introduce RightInspector shell; PropertiesPanel becomes shim
- New RightInspector.tsx is the right-side inspector entry point. Reads
  selectedIds[0], discriminates by entity type, and mounts the matching
  *Inspector. Bulk-actions branch (selectedIds.length > 1) inlined verbatim
  per D-05. Returns null on empty selection per D-01.
- App.tsx imports RightInspector and replaces both PropertiesPanel mounts
  (2D/split branch L282; 3D branch L314). AnimatePresence keys preserved.
- PropertiesPanel.tsx kept as a thin compatibility shim:
    * Delegates to RightInspector when something is selected.
    * Preserves the empty-state Room properties block (Phase 62 D-04 /
      polygonArea AREA row) so PropertiesPanel.area.test.tsx C1 still
      passes. App.tsx never hits this branch — the mount is gated on
      selectedIds.length > 0.
    * Doc-comment satisfies Phase 33 file-level contract test that pegs
      RotationPresetChips + rotateProduct strings to this path.
- TS clean; build clean; 996/996 vitest pass.
- Pre-existing failures in tests/SaveIndicator.test.tsx +
  tests/SidebarProductPicker.test.tsx are unrelated (verified on
  pre-Task-4 worktree).
