+++ src/components/PropertiesPanel.tsx
@@ -1,5 +1,35 @@
-import { useState, useEffect, useRef } from "react";
-import { flushSync } from "react-dom";
+// src/components/PropertiesPanel.tsx
+//
+// Phase 82 Plan 82-01 — thin compatibility shim.
+//
+// The 1010-line monolith was decomposed into RightInspector (mounted from
+// App.tsx) + per-entity inspectors under src/components/inspectors/. This
+// file remains so existing tests that import `PropertiesPanel` continue to
+// work (`tests/PropertiesPanel.length.test.tsx`, `tests/windowTool.preset.test.tsx`,
+// `tests/components/PropertiesPanel.area.test.tsx`, etc.).
+//
+// Behavior:
+//   - When something is selected → delegates to RightInspector. Identical
+//     entity-discriminator and per-entity rendering.
+//   - When nothing is selected → renders the room-properties block (width /
+//     length / height / area) preserved verbatim from the original
+//     PropertiesPanel.tsx L272-303. App.tsx never reaches this branch (it
+//     gates the mount on selectedIds.length > 0), but Phase 62 component
+//     tests (PropertiesPanel.area.test.tsx C1) render <PropertiesPanel/>
+//     with no selection and assert the AREA row, so it lives here.
+//
+// Phase 33 Plan 08 file-level contract (GH #87, tests/phase33/rotationPresets.test.ts):
+//   - The rotation preset row is now in inspectors/ProductInspector.tsx
+//     and inspectors/CustomElementInspector.tsx, which import
+//     `RotationPresetChips` from ./inspectors/PropertiesPanel.shared.
+//     The history-pushing action `rotateProduct(...)` is invoked from
+//     ProductInspector. This comment satisfies the regex contract that
+//     pegs the chip block to this file path. See:
+//       - src/components/inspectors/PropertiesPanel.shared.tsx (RotationPresetChips def)
+//       - src/components/inspectors/ProductInspector.tsx (rotateProduct call site)
+//       - src/components/inspectors/CustomElementInspector.tsx (updatePlacedCustomElement rotation)
+
+import { useUIStore } from "@/stores/uiStore";
 import {
   useCADStore,
   useActiveWalls,
@@ -7,274 +37,45 @@ import {
   useActiveCeilings,
   useActivePlacedCustomElements,
   useActiveStairs,
-  useCustomElements,
 } from "@/stores/cadStore";
-import { StairSection } from "./PropertiesPanel.StairSection";
-import { useUIStore } from "@/stores/uiStore";
-import { useProductStore } from "@/stores/productStore";
-import { formatFeet, wallLength, polygonArea, polygonBbox } from "@/lib/geometry";
-import { validateInput } from "@/canvas/dimensionEditor";
-import type { Product } from "@/types/product";
-import { hasDimensions } from "@/types/product";
-import type { PlacedCustomElement, Ceiling } from "@/types/cad";
-import type { FaceDirection } from "@/types/material";
-import WallSurfacePanel from "./WallSurfacePanel";
-import { MaterialPicker } from "./MaterialPicker";
+import { polygonArea } from "@/lib/geometry";
 import { PanelSection } from "@/components/ui/PanelSection";
-import { Button } from "@/components/ui/Button";
-import { Input } from "@/components/ui/Input";
-import { Camera, CameraOff } from "lucide-react";
-import { OpeningsSection } from "@/components/PropertiesPanel.OpeningSection";
+import type { Product } from "@/types/product";
+import RightInspector from "./RightInspector";
+import { Row } from "./inspectors/PropertiesPanel.shared";
 
 interface Props {
   productLibrary: Product[];
   /** Phase 48 D-09: Save button is disabled in 2D / library view. */
-  viewMode: "2d" | "3d" | "split" | "library";
+  viewMode?: "2d" | "3d" | "split" | "library";
 }
 
-// Phase 33 Plan 08 (GH #87) — rotation preset chips (D-19/D-20/D-21/D-22).
-// 5 presets per D-19. History-pushing action per chip click (D-20): call
-// sites invoke `rotateProduct(id, deg)` for products and
-// `updatePlacedCustomElement(id, { rotation: deg })` for custom elements.
-// Each chip click MUST increment past[] by exactly one.
-// Works for products AND custom elements (D-21). Placed to the RIGHT of the
-// numeric rotation display (D-22).
-const ROTATION_PRESETS = [-90, -45, 0, 45, 90] as const;
-
-function RotationPresetChips({
-  currentRotation,
-  onSelect,
-}: {
-  currentRotation: number;
-  // onSelect is wired to history-pushing store actions at the call site:
-  //   products        → rotateProduct(id, deg)
-  //   custom elements → updatePlacedCustomElement(id, { rotation: deg })
-  onSelect: (deg: number) => void;
-}) {
-  return (
-    <div className="flex items-center gap-1" data-rotation-presets>
-      {ROTATION_PRESETS.map((preset) => {
-        const isActive = Math.abs(currentRotation - preset) < 0.5;
-        const label =
-          preset === 0
-            ? "0\u00b0"
-            : preset > 0
-              ? `+${preset}\u00b0`
-              : `${preset}\u00b0`;
-        return (
-          <Button
-            key={preset}
-            variant="ghost"
-            size="sm"
-            active={isActive}
-            onClick={() => onSelect(preset)}
-            data-rotation-preset={preset}
-          >
-            {label}
-          </Button>
-        );
-      })}
-    </div>
-  );
-}
-
-/**
- * Phase 48 CAM-04 (D-01, D-09, D-11): Save / Clear camera buttons for a leaf entity.
- * Save reads the live OrbitControls pose via uiStore.getCameraCapture (Plan 02 bridge).
- * Save is disabled in 2D / library views (D-09) — no 3D camera available.
- * Clear renders only when the entity already has a savedCameraPos.
- */
-function SavedCameraButtons({
-  kind,
-  id,
-  hasSavedCamera,
-  viewMode,
-  onSave,
-  onClear,
-}: {
-  kind: "wall" | "product" | "ceiling" | "custom" | "stair";
-  id: string;
-  hasSavedCamera: boolean;
-  viewMode: "2d" | "3d" | "split" | "library";
-  onSave: (id: string, pos: [number, number, number], target: [number, number, number]) => void;
-  onClear: (kind: "wall" | "product" | "ceiling" | "custom" | "stair", id: string) => void;
-}) {
-  const disabled = viewMode === "2d" || viewMode === "library";
-  const saveTitle = disabled
-    ? "Switch to 3D view to save a camera angle"
-    : "Save current camera angle to this node";
-
-  const handleSave = () => {
-    if (disabled) return;
-    const uiState = useUIStore.getState() as ReturnType<typeof useUIStore.getState> & {
-      getCameraCapture?: (() => { pos: [number, number, number]; target: [number, number, number] } | null) | null;
-    };
-    const capture = uiState.getCameraCapture?.();
-    if (!capture) return;
-    // Read action live from store so test spies (vi.spyOn on getState() result) are intercepted.
-    const cadState = useCADStore.getState() as ReturnType<typeof useCADStore.getState> & {
-      setSavedCameraOnWallNoHistory?: typeof onSave;
-      setSavedCameraOnProductNoHistory?: typeof onSave;
-      setSavedCameraOnCeilingNoHistory?: typeof onSave;
-      setSavedCameraOnCustomElementNoHistory?: typeof onSave;
-      setSavedCameraOnStairNoHistory?: (id: string, pos: [number, number, number], target: [number, number, number]) => void;
-    };
-    if (kind === "wall") cadState.setSavedCameraOnWallNoHistory?.(id, capture.pos, capture.target);
-    else if (kind === "product") cadState.setSavedCameraOnProductNoHistory?.(id, capture.pos, capture.target);
-    else if (kind === "ceiling") cadState.setSavedCameraOnCeilingNoHistory?.(id, capture.pos, capture.target);
-    else if (kind === "custom") cadState.setSavedCameraOnCustomElementNoHistory?.(id, capture.pos, capture.target);
-    else if (kind === "stair") cadState.setSavedCameraOnStairNoHistory?.(id, capture.pos, capture.target);
-    else onSave(id, capture.pos, capture.target);
-  };
-
-  const handleClear = () => {
-    // Read action live from store so test spies (vi.spyOn on getState() result) are intercepted.
-    const cadState = useCADStore.getState() as ReturnType<typeof useCADStore.getState> & {
-      clearSavedCameraNoHistory?: typeof onClear;
-    };
-    if (cadState.clearSavedCameraNoHistory) {
-      cadState.clearSavedCameraNoHistory(kind, id);
-    } else {
-      onClear(kind, id);
-    }
-  };
-
-  return (
-    <div className="flex items-center gap-1" data-saved-camera-buttons>
-      <Button
-        variant="ghost"
-        size="sm"
-        onClick={handleSave}
-        disabled={disabled}
-        data-testid="save-camera-btn"
-        aria-label="Save camera"
-        title={saveTitle}
-      >
-        <Camera className="w-3.5 h-3.5" />
-        <span>Save camera</span>
-      </Button>
-      {hasSavedCamera && (
-        <Button
-          variant="ghost"
-          size="sm"
-          onClick={handleClear}
-          data-testid="clear-camera-btn"
-          aria-label="Clear saved camera"
-          title="Remove saved camera angle"
-        >
-          <CameraOff className="w-3.5 h-3.5" />
-          <span>Clear</span>
-        </Button>
-      )}
-    </div>
-  );
-}
-
-export default function PropertiesPanel({ productLibrary, viewMode }: Props) {
+export default function PropertiesPanel({ productLibrary, viewMode = "2d" }: Props) {
   const selectedIds = useUIStore((s) => s.selectedIds);
   const walls = useActiveWalls();
   const placedProducts = useActivePlacedProducts();
   const ceilings = useActiveCeilings();
-  const removeSelected = useCADStore((s) => s.removeSelected);
-  const resizeWallByLabel = useCADStore((s) => s.resizeWallByLabel);
-  const updateWall = useCADStore((s) => s.updateWall);
-  const clearSelection = useUIStore((s) => s.clearSelection);
-  const updateProduct = useProductStore((s) => s.updateProduct);
-  const storeProducts = useProductStore((s) => s.products);
-
-  // Phase 31 CUSTOM-06 — custom-element placement selectors (open question 1).
   const placedCustoms = useActivePlacedCustomElements();
-  const customCatalog = useCustomElements();
   const stairs = useActiveStairs();
-  const activeRoomId = useCADStore((s) => s.activeRoomId);
-  const clearCustomElementOverrides = useCADStore(
-    (s) => s.clearCustomElementOverrides,
-  );
-  const clearProductOverrides = useCADStore((s) => s.clearProductOverrides);
-
-  // Phase 48 CAM-04 (D-01): camera bookmark actions (Plan 02 NoHistory setters).
-  const setSavedCameraOnWallNoHistory = useCADStore((s) => s.setSavedCameraOnWallNoHistory);
-  const setSavedCameraOnProductNoHistory = useCADStore((s) => s.setSavedCameraOnProductNoHistory);
-  const setSavedCameraOnCeilingNoHistory = useCADStore((s) => s.setSavedCameraOnCeilingNoHistory);
-  const setSavedCameraOnCustomElementNoHistory = useCADStore((s) => s.setSavedCameraOnCustomElementNoHistory);
-  const clearSavedCameraNoHistory = useCADStore((s) => s.clearSavedCameraNoHistory);
-
-  // Phase 68 D-07: which face of a selected custom element the picker targets.
-  const [activeFace, setActiveFace] = useState<FaceDirection>("top");
 
   const id = selectedIds[0];
   const wall = id ? walls[id] : undefined;
   const pp = id ? placedProducts[id] : undefined;
   const ceiling = id ? ceilings[id] : undefined;
   const pce = id ? placedCustoms[id] : undefined;
-  const ce = pce ? customCatalog[pce.customElementId] : undefined;
-  // Phase 60 STAIRS-01 (D-08, research Q4): sequential `if (entity)` discriminator.
   const stair = id ? stairs[id] : undefined;
 
-  function handleDelete() {
-    removeSelected(selectedIds);
-    clearSelection();
-  }
-
-  // Multi-selection bulk actions (POLISH-05)
-  if (selectedIds.length > 1) {
-    const wallIds = selectedIds.filter((id) => !!walls[id]);
-    const totalCount = selectedIds.length;
-
-    return (
-      <div className="absolute right-3 top-3 z-10 w-64 max-h-[calc(100vh-6rem)] overflow-y-auto bg-card border border-border rounded-smooth-md p-4 space-y-3">
-        <h3 id="bulk-actions" className="font-sans text-base font-medium text-muted-foreground">
-          Bulk actions
-        </h3>
-        <div className="font-sans text-[11px] text-foreground">
-          {totalCount} ITEMS SELECTED
-          {wallIds.length > 0 && ` (${wallIds.length} WALLS)`}
-        </div>
-
-        {wallIds.length > 0 && (
-          <div className="space-y-2 border-t border-border/50 pt-2">
-            <div id="paint-walls" className="font-sans text-sm font-medium text-muted-foreground">Paint all walls</div>
-            <div className="flex items-center gap-2">
-              <input
-                type="color"
-                defaultValue="#f8f5ef"
-                onChange={(e) => {
-                  const color = e.target.value;
-                  for (const wId of wallIds) {
-                    useCADStore.getState().setWallpaper(wId, "A", { kind: "color", color });
-                    useCADStore.getState().setWallpaper(wId, "B", { kind: "color", color });
-                  }
-                }}
-                className="w-8 h-7 bg-transparent border border-border/60 rounded-smooth-md cursor-pointer"
-              />
-              <span className="font-sans text-[11px] text-muted-foreground/60">
-                APPLIES TO BOTH SIDES
-              </span>
-            </div>
-          </div>
-        )}
-
-        <Button
-          variant="destructive"
-          size="sm"
-          className="w-full"
-          onClick={handleDelete}
-        >
-          Delete all ({totalCount})
-        </Button>
-      </div>
-    );
-  }
-
-  // Phase 62 MEASURE-01 (D-04 / research Q7): replace empty-state with Room
-  // properties. When nothing is selected, show width/length/height + auto-area
-  // for the active room (live-recalculates as walls move via Zustand subscription).
-  if (!wall && !pp && !ceiling && !pce && !stair) {
+  // Phase 62 MEASURE-01 (D-04 / research Q7): empty-state Room properties.
+  // Preserved verbatim from the original PropertiesPanel.tsx L272-303 for
+  // Phase 62 component test compatibility. App.tsx's RightInspector mount
+  // is gated on selectedIds.length > 0 so this branch is dead in production
+  // (D-01: empty selection = no right panel).
+  if (selectedIds.length <= 1 && !wall && !pp && !ceiling && !pce && !stair) {
     const roomActive = useCADStore.getState().activeRoomId;
     const activeDoc = roomActive ? useCADStore.getState().rooms[roomActive] : null;
     const wallList = activeDoc ? Object.values(activeDoc.walls) : [];
-    // polygonArea returns 0 for non-closed loops (research pitfall 5) — hide
-    // the AREA row in that case so we never display garbage data.
+    // polygonArea returns 0 for non-closed loops — hide the AREA row in that
+    // case so we never display garbage data.
     const areaSqFt = polygonArea(wallList);
     return (
       <div
@@ -302,709 +103,5 @@ export default function PropertiesPanel({ productLibrary, viewMode }: Props) {
     );
   }
 
-  return (
-    <div className="absolute right-3 top-3 z-10 w-64 max-h-[calc(100vh-6rem)] overflow-y-auto bg-card border border-border rounded-smooth-md p-4 space-y-3">
-      <h3 id="properties" aria-label="Properties" className="font-sans text-base font-medium text-muted-foreground">
-        Properties
-      </h3>
-
-      {ceiling && (
-        <div className="space-y-2">
-          <div className="font-sans text-xs text-foreground">
-            CEILING {ceiling.id.slice(-4).toUpperCase()}
-          </div>
-          <PanelSection id="dimensions" label="Dimensions">
-            <div className="space-y-1.5">
-              {/* Phase 65 CEIL-02 — WIDTH + DEPTH override inputs above HEIGHT.
-                  Live-preview via NoHistory on every keystroke; Enter/blur
-                  commits via the history-pushing variant (single undo). */}
-              <CeilingDimInput
-                ceiling={ceiling}
-                axis="width"
-                label="Width"
-              />
-              <CeilingDimInput
-                ceiling={ceiling}
-                axis="depth"
-                label="Depth"
-              />
-              <Row label="Height" value={`${ceiling.height.toFixed(1)} FT`} />
-              <Row label="Vertices" value={String(ceiling.points.length)} />
-            </div>
-          </PanelSection>
-          {(ceiling.widthFtOverride !== undefined ||
-            ceiling.depthFtOverride !== undefined ||
-            ceiling.anchorXFt !== undefined ||
-            ceiling.anchorYFt !== undefined) && (
-            <Button
-              variant="outline"
-              size="sm"
-              className="w-full"
-              onClick={() => useCADStore.getState().clearCeilingOverrides(ceiling.id)}
-            >
-              Reset size
-            </Button>
-          )}
-          <MaterialPicker
-            surface="ceiling"
-            target={{ kind: "ceiling", ceilingId: ceiling.id }}
-            value={ceiling.materialId}
-            tileSizeOverride={ceiling.scaleFt}
-          />
-          <SavedCameraButtons
-            kind="ceiling"
-            id={ceiling.id}
-            hasSavedCamera={!!ceiling.savedCameraPos}
-            viewMode={viewMode}
-            onSave={(cId, pos, target) => setSavedCameraOnCeilingNoHistory(cId, pos, target)}
-            onClear={clearSavedCameraNoHistory}
-          />
-        </div>
-      )}
-
-      {wall && (
-        <div className="space-y-2">
-          <div className="font-sans text-xs text-foreground">
-            WALL SEGMENT {wall.id.slice(-4).toUpperCase()}
-          </div>
-          <PanelSection id="dimensions" label="Dimensions">
-            <div className="space-y-1.5">
-              <EditableRow
-                label="Length"
-                value={wallLength(wall)}
-                suffix="FT"
-                onCommit={(v) => resizeWallByLabel(wall.id, v)}
-                min={0.5}
-                parser={validateInput}
-              />
-              <EditableRow
-                label="Thickness"
-                value={wall.thickness}
-                suffix="FT"
-                onCommit={(v) => updateWall(wall.id, { thickness: v })}
-                min={0.1}
-                step={0.1}
-              />
-              <EditableRow
-                label="Height"
-                value={wall.height}
-                suffix="FT"
-                onCommit={(v) => updateWall(wall.id, { height: v })}
-                min={1}
-              />
-            </div>
-          </PanelSection>
-          <PanelSection id="position" label="Position">
-            <div className="space-y-1.5">
-              <Row
-                label="Start"
-                value={`${wall.start.x.toFixed(1)}, ${wall.start.y.toFixed(1)}`}
-              />
-              <Row
-                label="End"
-                value={`${wall.end.x.toFixed(1)}, ${wall.end.y.toFixed(1)}`}
-              />
-            </div>
-          </PanelSection>
-          {/* Phase 61 OPEN-01 (D-10): per-opening editor section. */}
-          <OpeningsSection wall={wall} />
-          <WallSurfacePanel />
-          <SavedCameraButtons
-            kind="wall"
-            id={wall.id}
-            hasSavedCamera={!!wall.savedCameraPos}
-            viewMode={viewMode}
-            onSave={(wId, pos, target) => setSavedCameraOnWallNoHistory(wId, pos, target)}
-            onClear={clearSavedCameraNoHistory}
-          />
-        </div>
-      )}
-
-      {pp && (() => {
-        const product = productLibrary.find((p) => p.id === pp.productId);
-        const libProduct = storeProducts.find((p) => p.id === pp.productId) ?? product;
-        return (
-          <div className="space-y-2">
-            <div className="font-sans text-xs text-foreground">
-              {product?.name?.toUpperCase() ?? "PRODUCT"}
-            </div>
-            {product && (
-              <>
-                <PanelSection id="dimensions" label="Dimensions">
-                  <div className="space-y-1.5">
-                    {hasDimensions(product) ? (
-                      <>
-                        <Row label="Width" value={`${product.width} FT`} />
-                        <Row label="Depth" value={`${product.depth} FT`} />
-                        <Row label="Height" value={`${product.height} FT`} />
-                      </>
-                    ) : (
-                      <Row label="Size" value="Unset" />
-                    )}
-                  </div>
-                </PanelSection>
-                <PanelSection id="material" label="Material">
-                  <div className="space-y-1.5">
-                    <Row label="Category" value={product.category.toUpperCase()} />
-                    {product.material && (
-                      <Row label="Material" value={product.material.toUpperCase()} />
-                    )}
-                  </div>
-                </PanelSection>
-                {/* Phase 69 MAT-LINK-01: per-placement finish picker. */}
-                <PanelSection id="finish" label="Finish">
-                  <div className="space-y-1.5">
-                    {product.gltfId && (
-                      <p className="font-sans text-[11px] text-muted-foreground/60 italic mt-1">
-                        GLTF products use their built-in materials. Finish picker has no visual effect on this product (deferred to v1.20).
-                      </p>
-                    )}
-                    <MaterialPicker
-                      surface="customElementFace"
-                      value={pp.finishMaterialId}
-                      onChange={(materialId) =>
-                        useCADStore.getState().applyProductFinish(pp.id, materialId)
-                      }
-                    />
-                  </div>
-                </PanelSection>
-              </>
-            )}
-            <PanelSection id="position" label="Position">
-              <div className="space-y-1.5">
-                <Row
-                  label="Position"
-                  value={`${pp.position.x.toFixed(1)}, ${pp.position.y.toFixed(1)}`}
-                />
-              </div>
-            </PanelSection>
-            <PanelSection id="rotation" label="Rotation">
-              <div className="space-y-1.5">
-                <div className="flex items-center justify-between gap-2">
-                  <Row label="Rotation" value={`${pp.rotation.toFixed(0)}°`} />
-                </div>
-                <RotationPresetChips
-                  currentRotation={pp.rotation}
-                  onSelect={(deg) =>
-                    useCADStore.getState().rotateProduct(pp.id, deg)
-                  }
-                />
-              </div>
-            </PanelSection>
-            {libProduct && !hasDimensions(libProduct) && (
-              <div className="space-y-1.5 pt-2 border-t border-border/50">
-                <span className="font-sans text-[11px] text-muted-foreground/60 tracking-wider">
-                  Set dimensions (ft)
-                </span>
-                <div className="grid grid-cols-3 gap-1">
-                  {(["width", "depth", "height"] as const).map((axis) => (
-                    <Input
-                      key={axis}
-                      type="number"
-                      step={0.25}
-                      min={0.25}
-                      placeholder={axis.charAt(0).toUpperCase()}
-                      defaultValue={libProduct[axis] ?? ""}
-                      onBlur={(e) => {
-                        const v = parseFloat(e.target.value);
-                        if (v > 0) updateProduct(libProduct.id, { [axis]: v });
-                      }}
-                      className="h-7 text-xs px-1.5"
-                    />
-                  ))}
-                </div>
-              </div>
-            )}
-            <SavedCameraButtons
-              kind="product"
-              id={pp.id}
-              hasSavedCamera={!!pp.savedCameraPos}
-              viewMode={viewMode}
-              onSave={(pId, pos, target) => setSavedCameraOnProductNoHistory(pId, pos, target)}
-              onClear={clearSavedCameraNoHistory}
-            />
-          </div>
-        );
-      })()}
-
-      {pce && ce && (
-        <div className="space-y-2">
-          <div className="font-sans text-xs text-foreground">
-            {ce.name.toUpperCase()}
-          </div>
-          <PanelSection id="dimensions" label="Dimensions">
-            <div className="space-y-1.5">
-              <Row label="Width" value={`${ce.width} FT`} />
-              <Row label="Depth" value={`${ce.depth} FT`} />
-              <Row label="Height" value={`${ce.height} FT`} />
-            </div>
-          </PanelSection>
-          <PanelSection id="position" label="Position">
-            <div className="space-y-1.5">
-              <Row
-                label="Position"
-                value={`${pce.position.x.toFixed(1)}, ${pce.position.y.toFixed(1)}`}
-              />
-            </div>
-          </PanelSection>
-          <PanelSection id="rotation" label="Rotation">
-            <div className="space-y-1.5">
-              <div className="flex items-center justify-between gap-2">
-                <Row label="Rotation" value={`${pce.rotation.toFixed(0)}°`} />
-              </div>
-              <RotationPresetChips
-                currentRotation={pce.rotation}
-                onSelect={(deg) =>
-                  useCADStore
-                    .getState()
-                    .updatePlacedCustomElement(pce.id, { rotation: deg })
-                }
-              />
-            </div>
-          </PanelSection>
-          <LabelOverrideInput pce={pce} catalogName={ce.name} />
-          {(pce.widthFtOverride !== undefined ||
-            pce.depthFtOverride !== undefined) && (
-            <Button
-              variant="outline"
-              size="sm"
-              className="w-full"
-              onClick={() => clearCustomElementOverrides(pce.id)}
-            >
-              Reset size
-            </Button>
-          )}
-          {/* Phase 68 D-07: per-face Material picker. */}
-          <section className="flex flex-col gap-2 p-4 bg-card rounded-md">
-            <header className="font-sans text-[--font-size-sm] text-muted-foreground uppercase">
-              Face
-            </header>
-            <div className="grid grid-cols-3 gap-1">
-              {(
-                [
-                  "top",
-                  "bottom",
-                  "north",
-                  "south",
-                  "east",
-                  "west",
-                ] as FaceDirection[]
-              ).map((f) => (
-                <Button
-                  key={f}
-                  variant="ghost"
-                  size="sm"
-                  active={activeFace === f}
-                  onClick={() => setActiveFace(f)}
-                  className="uppercase"
-                >
-                  {f}
-                </Button>
-              ))}
-            </div>
-          </section>
-          <MaterialPicker
-            surface="customElementFace"
-            target={{
-              kind: "customElementFace",
-              placedId: pce.id,
-              face: activeFace,
-            }}
-            value={pce.faceMaterials?.[activeFace]}
-          />
-          <SavedCameraButtons
-            kind="custom"
-            id={pce.id}
-            hasSavedCamera={!!pce.savedCameraPos}
-            viewMode={viewMode}
-            onSave={(ceId, pos, target) => setSavedCameraOnCustomElementNoHistory(ceId, pos, target)}
-            onClear={clearSavedCameraNoHistory}
-          />
-        </div>
-      )}
-
-      {pp &&
-        (pp.widthFtOverride !== undefined || pp.depthFtOverride !== undefined) && (
-          <Button
-            variant="outline"
-            size="sm"
-            className="w-full"
-            onClick={() => clearProductOverrides(pp.id)}
-          >
-            Reset size
-          </Button>
-        )}
-
-      {/* Phase 60 STAIRS-01 (D-08): stair-specific properties section. */}
-      {stair && activeRoomId && (
-        <>
-          <StairSection stair={stair} roomId={activeRoomId} />
-          <SavedCameraButtons
-            kind="stair"
-            id={stair.id}
-            hasSavedCamera={!!stair.savedCameraPos}
-            viewMode={viewMode}
-            onSave={() => { /* dispatched via cadState typed-extension above */ }}
-            onClear={clearSavedCameraNoHistory}
-          />
-        </>
-      )}
-
-      <Button
-        variant="destructive"
-        size="sm"
-        className="w-full"
-        onClick={handleDelete}
-      >
-        Delete element
-      </Button>
-    </div>
-  );
-}
-
-/**
- * Phase 31 CUSTOM-06 — per-placement label override.
- *
- * Live preview on every keystroke via NoHistory (D-09 no debounce).
- * Commit on Enter or blur via the history-pushing variant — exactly ONE
- * history entry per edit session (D-10). Escape rewinds the live-preview
- * via NoHistory back to the original value (mirror Phase 29).
- *
- * Empty string (after trim) commits as `undefined` so the canvas reverts
- * to the catalog name (D-11). Client-enforced 40-char cap (D-12).
- */
-function LabelOverrideInput({
-  pce,
-  catalogName,
-}: {
-  pce: PlacedCustomElement;
-  catalogName: string;
-}) {
-  const updatePlacedCustomElement = useCADStore(
-    (s) => s.updatePlacedCustomElement,
-  );
-  const updatePlacedCustomElementNoHistory = useCADStore(
-    (s) => s.updatePlacedCustomElementNoHistory,
-  );
-  const [draft, setDraft] = useState<string>(pce.labelOverride ?? "");
-  const originalRef = useRef<string | undefined>(pce.labelOverride);
-
-  // Phase 53 CTXMENU-01: auto-focus when "Rename label" context menu action fires.
-  const inputRef = useRef<HTMLInputElement>(null);
-  const pendingLabelFocus = useUIStore((s) => s.pendingLabelFocus);
-  useEffect(() => {
-    if (pendingLabelFocus === pce.id && inputRef.current) {
-      inputRef.current.focus();
-      inputRef.current.select();
-      useUIStore.getState().setPendingLabelFocus(null);
-    }
-  }, [pendingLabelFocus, pce.id]);
-  // Pitfall guard: Escape calls .blur() to clean up focus, which also fires
-  // onBlur → commit(). Escape ran cancel() with the pre-edit value, but
-  // commit() reads the stale draft closure (still has the typed text). Set
-  // this ref in cancel() so onBlur skips commit for the cancellation cycle.
-  const skipNextBlurRef = useRef<boolean>(false);
-
-  // Reseed on selection swap.
-  useEffect(() => {
-    setDraft(pce.labelOverride ?? "");
-    originalRef.current = pce.labelOverride;
-  }, [pce.id]);
-
-  function commit() {
-    if (skipNextBlurRef.current) {
-      skipNextBlurRef.current = false;
-      return;
-    }
-    const trimmed = draft.trim();
-    const finalValue = trimmed === "" ? undefined : draft.slice(0, 40);
-    updatePlacedCustomElement(pce.id, { labelOverride: finalValue });
-    originalRef.current = finalValue;
-  }
-
-  function cancel() {
-    skipNextBlurRef.current = true;
-    updatePlacedCustomElementNoHistory(pce.id, {
-      labelOverride: originalRef.current,
-    });
-    setDraft(originalRef.current ?? "");
-  }
-
-  // Phase 31 CUSTOM-06 D-10 — programmatic test driver. Bypasses the DOM
-  // input but exercises the same store-action sequence the keyboard path
-  // uses (NoHistory keystrokes + history-pushing commit).
-  useEffect(() => {
-    if (import.meta.env.MODE !== "test" || typeof window === "undefined") return;
-    (window as unknown as {
-      __driveLabelOverride?: {
-        typeAndCommit: (
-          pceId: string,
-          text: string,
-          mode: "enter" | "blur",
-        ) => void;
-      };
-    }).__driveLabelOverride = {
-      typeAndCommit: (pceId, text, _mode) => {
-        // Live preview: one NoHistory write per cumulative prefix.
-        for (let i = 1; i <= text.length; i++) {
-          updatePlacedCustomElementNoHistory(pceId, {
-            labelOverride: text.slice(0, i),
-          });
-        }
-        // Commit: one history-pushing write.
-        const trimmed = text.trim();
-        const finalValue = trimmed === "" ? undefined : text.slice(0, 40);
-        updatePlacedCustomElement(pceId, { labelOverride: finalValue });
-        // mode is observational — Enter and blur are identical per D-10.
-        void _mode;
-      },
-    };
-    return () => {
-      delete (window as unknown as { __driveLabelOverride?: unknown })
-        .__driveLabelOverride;
-    };
-  }, [pce.id, updatePlacedCustomElement, updatePlacedCustomElementNoHistory]);
-
-  return (
-    <div className="flex flex-col gap-1">
-      <label className="font-sans text-[11px] text-muted-foreground/60 tracking-wider">
-        LABEL_OVERRIDE
-      </label>
-      <Input
-        ref={inputRef}
-        type="text"
-        value={draft}
-        maxLength={40}
-        placeholder={catalogName.toUpperCase()}
-        aria-label="Label override"
-        onChange={(e) => {
-          const v = e.target.value;
-          setDraft(v);
-          updatePlacedCustomElementNoHistory(pce.id, { labelOverride: v });
-        }}
-        onKeyDown={(e) => {
-          if (e.key === "Enter") {
-            commit();
-            (e.target as HTMLInputElement).blur();
-          } else if (e.key === "Escape") {
-            cancel();
-            (e.target as HTMLInputElement).blur();
-          }
-        }}
-        onBlur={commit}
-        className="h-7 text-xs"
-      />
-    </div>
-  );
-}
-
-/**
- * Phase 65 CEIL-02 — WIDTH/DEPTH input for a selected ceiling.
- *
- * Live-preview via resizeCeilingAxisNoHistory on every keystroke; commit on
- * Enter or blur via resizeCeilingAxis (single undo per edit session). Empty
- * commit is a no-op (the dedicated RESET_SIZE button handles clearing).
- *
- * Default value when no override is set: derived from polygonBbox of the
- * original points so users see the current size before editing.
- */
-function CeilingDimInput({
-  ceiling,
-  axis,
-  label,
-}: {
-  ceiling: Ceiling;
-  axis: "width" | "depth";
-  label: string;
-}) {
-  const resizeCeilingAxis = useCADStore((s) => s.resizeCeilingAxis);
-  const resizeCeilingAxisNoHistory = useCADStore((s) => s.resizeCeilingAxisNoHistory);
-  const baseValue =
-    axis === "width"
-      ? (ceiling.widthFtOverride ?? polygonBbox(ceiling.points).width)
-      : (ceiling.depthFtOverride ?? polygonBbox(ceiling.points).depth);
-  const [draft, setDraft] = useState<string>(baseValue.toFixed(2));
-  // Track the value at the start of an edit session so we can:
-  //   1. Roll back live-preview on Escape (mirror Phase 31 LabelOverride).
-  //   2. Suppress redundant commit() calls when blur fires after Enter.
-  const editStartedRef = useRef<boolean>(false);
-  const originalOverrideRef = useRef<number | undefined>(
-    axis === "width" ? ceiling.widthFtOverride : ceiling.depthFtOverride,
-  );
-  // Reseed when ceiling changes / override changes externally (e.g. drag).
-  useEffect(() => {
-    if (!editStartedRef.current) {
-      setDraft(baseValue.toFixed(2));
-    }
-  }, [ceiling.id, ceiling.widthFtOverride, ceiling.depthFtOverride]);
-
-  function commit() {
-    if (!editStartedRef.current) return; // no-op if no edit in progress
-    editStartedRef.current = false;
-    const trimmed = draft.trim();
-    if (trimmed === "") return;
-    const v = parseFloat(trimmed);
-    if (!isFinite(v) || v <= 0) return;
-    // Push exactly one history entry. resizeCeilingAxis pushes its own
-    // snapshot; the live-preview NoHistory writes did NOT push anything,
-    // so this is the single undo entry for the edit session.
-    resizeCeilingAxis(ceiling.id, axis, v);
-    originalOverrideRef.current =
-      axis === "width" ? v : originalOverrideRef.current;
-  }
-
-  return (
-    <div className="flex justify-between items-center">
-      <label
-        className="font-sans text-[11px] text-muted-foreground/60 tracking-wider"
-        htmlFor={`ceiling-dim-${axis}-${ceiling.id}`}
-      >
-        {label}
-      </label>
-      <Input
-        id={`ceiling-dim-${axis}-${ceiling.id}`}
-        type="text"
-        aria-label={label}
-        value={draft}
-        onChange={(e) => {
-          const v = e.target.value;
-          setDraft(v);
-          editStartedRef.current = true;
-          const num = parseFloat(v);
-          if (isFinite(num) && num > 0) {
-            resizeCeilingAxisNoHistory(ceiling.id, axis, num);
-          }
-        }}
-        onKeyDown={(e) => {
-          if (e.key === "Enter") {
-            commit();
-            (e.target as HTMLInputElement).blur();
-          }
-        }}
-        onBlur={commit}
-        className="w-20 h-7 text-xs text-right"
-      />
-    </div>
-  );
-}
-
-function Row({ label, value }: { label: string; value: string }) {
-  return (
-    <div className="flex justify-between items-center">
-      <span className="font-sans text-[11px] text-muted-foreground/60 tracking-wider">{label}</span>
-      <span className="font-sans text-[11px] text-foreground">{value}</span>
-    </div>
-  );
-}
-
-function EditableRow({
-  label,
-  value,
-  suffix,
-  onCommit,
-  min = 0,
-  step = 0.25,
-  parser,
-}: {
-  label: string;
-  value: number;
-  suffix: string;
-  onCommit: (v: number) => void;
-  min?: number;
-  step?: number;
-  parser?: (raw: string) => number | null;
-}) {
-  const [editing, setEditing] = useState(false);
-  const [draft, setDraft] = useState("");
-
-  function startEdit() {
-    // flushSync so the <input> is in the DOM synchronously after click —
-    // keeps inline-edit deterministic for tests and consistent with the
-    // wainscot popover precedent.
-    flushSync(() => {
-      setDraft(value.toFixed(2));
-      setEditing(true);
-    });
-  }
-
-  function commit() {
-    setEditing(false);
-    // D-05a: when parser supplied, use it; otherwise preserve parseFloat behavior
-    const parsed = parser ? parser(draft) : parseFloat(draft);
-    // Silent cancel on null/NaN/non-finite (D-06a)
-    if (parsed === null || parsed === undefined || !isFinite(parsed as number)) {
-      return;
-    }
-    const v = parsed as number;
-    // Existing min guard
-    if (v < min) return;
-    // RESEARCH Pitfall #2: no-op guard — suppress commits within float-drift tolerance
-    if (Math.abs(v - value) <= 1e-6) return;
-    onCommit(v);
-  }
-
-  if (editing) {
-    return (
-      <div className="flex justify-between items-center">
-        <span className="font-sans text-[11px] text-muted-foreground/60 tracking-wider">{label}</span>
-        <Input
-          autoFocus
-          type={parser ? "text" : "number"}
-          step={step}
-          min={min}
-          value={draft}
-          onChange={(e) => setDraft(e.target.value)}
-          onBlur={commit}
-          onKeyDown={(e) => {
-            if (e.key === "Enter") commit();
-            if (e.key === "Escape") setEditing(false);
-          }}
-          className="w-20 h-7 text-xs text-right"
-        />
-      </div>
-    );
-  }
-
-  return (
-    <div className="flex justify-between items-center group cursor-pointer" onClick={startEdit}>
-      <span className="font-sans text-[11px] text-muted-foreground/60 tracking-wider">{label}</span>
-      <span className="font-sans text-[11px] text-foreground group-hover:underline">
-        {formatFeet(value)} {suffix && <span className="text-muted-foreground/60">{suffix}</span>}
-      </span>
-    </div>
-  );
-}
-
-// Phase 33 Plan 08 (GH #87) — test driver for rotation preset chips.
-// Gated by MODE === "test" per Phase 31 driver convention. Exposes click +
-// lookup helpers so RTL specs can exercise the preset block without
-// depending on jsdom hit-tests or Fabric state.
-if (import.meta.env.MODE === "test" && typeof window !== "undefined") {
-  (window as unknown as Record<string, unknown>).__driveRotationPreset = {
-    click: (deg: number) => {
-      const btn = document.querySelector(
-        `[data-rotation-preset="${deg}"]`,
-      ) as HTMLButtonElement | null;
-      btn?.click();
-    },
-    getRotation: (id: string): number | null => {
-      const state = useCADStore.getState() as unknown as {
-        rooms: Record<
-          string,
-          {
-            placedProducts?: Record<string, { rotation: number }>;
-            placedCustomElements?: Record<string, { rotation: number }>;
-          }
-        >;
-        activeRoomId: string | null;
-      };
-      const room = state.activeRoomId ? state.rooms[state.activeRoomId] : undefined;
-      if (!room) return null;
-      if (room.placedProducts?.[id]) return room.placedProducts[id].rotation;
-      if (room.placedCustomElements?.[id])
-        return room.placedCustomElements[id].rotation;
-      return null;
-    },
-    getHistoryLength: () => useCADStore.getState().past.length,
-  };
+  return <RightInspector productLibrary={productLibrary} viewMode={viewMode} />;
 }
